requirements.txt
# torch
numpy
librosa
pandas
feather-format
matplotlib
tqdm